+++ versions.tf
@@ -1,9 +1,9 @@
 terraform {
-  required_version = ">= 0.13.0"
+  required_version = ">= 1.1"
   required_providers {
     azurerm = {
       source  = "hashicorp/azurerm"
-      version = ">= 2.83"
+      version = "~> 3.0"
     }
     azurecaf = {
       source  = "aztfmod/azurecaf"