+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# unreleased
+
+Changed
+  * AZ-769: Update module to AzureRM V3
+
 # v5.5.0 - 2022-06-24
 
 Added